+++ src/BlaeckTCP.cpp
@@ -93,9 +93,6 @@ void BlaeckTCP::addSignal(String signalName, bool *value)
   Signals[_signalIndex].SignalName = signalName;
   Signals[_signalIndex].DataType = Blaeck_bool;
   Signals[_signalIndex].Address = value;
-#ifdef __AVR__
-  Signals[_signalIndex].UseFlashSignalName = false;
-#endif
   _signalIndex++;
 }
 
@@ -104,9 +101,6 @@ void BlaeckTCP::addSignal(String signalName, byte *value)
   Signals[_signalIndex].SignalName = signalName;
   Signals[_signalIndex].DataType = Blaeck_byte;
   Signals[_signalIndex].Address = value;
-#ifdef __AVR__
-  Signals[_signalIndex].UseFlashSignalName = false;
-#endif
   _signalIndex++;
 }
 
@@ -115,9 +109,6 @@ void BlaeckTCP::addSignal(String signalName, short *value)
   Signals[_signalIndex].SignalName = signalName;
   Signals[_signalIndex].DataType = Blaeck_short;
   Signals[_signalIndex].Address = value;
-#ifdef __AVR__
-  Signals[_signalIndex].UseFlashSignalName = false;
-#endif
   _signalIndex++;
 }
 
@@ -126,9 +117,6 @@ void BlaeckTCP::addSignal(String signalName, unsigned short *value)
   Signals[_signalIndex].SignalName = signalName;
   Signals[_signalIndex].DataType = Blaeck_ushort;
   Signals[_signalIndex].Address = value;
-#ifdef __AVR__
-  Signals[_signalIndex].UseFlashSignalName = false;
-#endif
   _signalIndex++;
 }
 
@@ -137,9 +125,6 @@ void BlaeckTCP::addSignal(String signalName, int *value)
   Signals[_signalIndex].SignalName = signalName;
   Signals[_signalIndex].DataType = Blaeck_int;
   Signals[_signalIndex].Address = value;
-#ifdef __AVR__
-  Signals[_signalIndex].UseFlashSignalName = false;
-#endif
   _signalIndex++;
 }
 
@@ -148,9 +133,6 @@ void BlaeckTCP::addSignal(String signalName, unsigned int *value)
   Signals[_signalIndex].SignalName = signalName;
   Signals[_signalIndex].DataType = Blaeck_uint;
   Signals[_signalIndex].Address = value;
-#ifdef __AVR__
-  Signals[_signalIndex].UseFlashSignalName = false;
-#endif
   _signalIndex++;
 }
 
@@ -159,9 +141,6 @@ void BlaeckTCP::addSignal(String signalName, long *value)
   Signals[_signalIndex].SignalName = signalName;
   Signals[_signalIndex].DataType = Blaeck_long;
   Signals[_signalIndex].Address = value;
-#ifdef __AVR__
-  Signals[_signalIndex].UseFlashSignalName = false;
-#endif
   _signalIndex++;
 }
 
@@ -170,9 +149,6 @@ void BlaeckTCP::addSignal(String signalName, unsigned long *value)
   Signals[_signalIndex].SignalName = signalName;
   Signals[_signalIndex].DataType = Blaeck_ulong;
   Signals[_signalIndex].Address = value;
-#ifdef __AVR__
-  Signals[_signalIndex].UseFlashSignalName = false;
-#endif
   _signalIndex++;
 }
 
@@ -181,9 +157,6 @@ void BlaeckTCP::addSignal(String signalName, float *value)
   Signals[_signalIndex].SignalName = signalName;
   Signals[_signalIndex].DataType = Blaeck_float;
   Signals[_signalIndex].Address = value;
-#ifdef __AVR__
-  Signals[_signalIndex].UseFlashSignalName = false;
-#endif
   _signalIndex++;
 }
 
@@ -194,7 +167,6 @@ void BlaeckTCP::addSignal(String signalName, double *value)
   /*On the Uno and other ATMEGA based boards, the double implementation occupies 4 bytes
   and is exactly the same as the float, with no gain in precision.*/
   Signals[_signalIndex].DataType = Blaeck_float;
-  Signals[_signalIndex].UseFlashSignalName = false;
 #else
   Signals[_signalIndex].DataType = Blaeck_double;
 #endif
@@ -206,108 +178,6 @@ void BlaeckTCP::deleteSignals()
 {
   _signalIndex = 0;
 }
-
-#ifdef __AVR__
-void BlaeckTCP::addSignal(PGM_P const *signalNameTable, int signalNameIndex, bool *value)
-{
-  Signals[_signalIndex].SignalNameTable = signalNameTable;
-  Signals[_signalIndex].SignalNameIndex = signalNameIndex;
-  Signals[_signalIndex].UseFlashSignalName = true;
-  Signals[_signalIndex].DataType = Blaeck_bool;
-  Signals[_signalIndex].Address = value;
-  _signalIndex++;
-}
-
-void BlaeckTCP::addSignal(PGM_P const *signalNameTable, int signalNameIndex, byte *value)
-{
-  Signals[_signalIndex].SignalNameTable = signalNameTable;
-  Signals[_signalIndex].SignalNameIndex = signalNameIndex;
-  Signals[_signalIndex].UseFlashSignalName = true;
-  Signals[_signalIndex].DataType = Blaeck_byte;
-  Signals[_signalIndex].Address = value;
-  _signalIndex++;
-}
-
-void BlaeckTCP::addSignal(PGM_P const *signalNameTable, int signalNameIndex, short *value)
-{
-  Signals[_signalIndex].SignalNameTable = signalNameTable;
-  Signals[_signalIndex].SignalNameIndex = signalNameIndex;
-  Signals[_signalIndex].UseFlashSignalName = true;
-  Signals[_signalIndex].DataType = Blaeck_short;
-  Signals[_signalIndex].Address = value;
-  _signalIndex++;
-}
-
-void BlaeckTCP::addSignal(PGM_P const *signalNameTable, int signalNameIndex, unsigned short *value)
-{
-  Signals[_signalIndex].SignalNameTable = signalNameTable;
-  Signals[_signalIndex].SignalNameIndex = signalNameIndex;
-  Signals[_signalIndex].UseFlashSignalName = true;
-  Signals[_signalIndex].DataType = Blaeck_ushort;
-  Signals[_signalIndex].Address = value;
-  _signalIndex++;
-}
-
-void BlaeckTCP::addSignal(PGM_P const *signalNameTable, int signalNameIndex, int *value)
-{
-  Signals[_signalIndex].SignalNameTable = signalNameTable;
-  Signals[_signalIndex].SignalNameIndex = signalNameIndex;
-  Signals[_signalIndex].UseFlashSignalName = true;
-  Signals[_signalIndex].DataType = Blaeck_int;
-  Signals[_signalIndex].Address = value;
-  _signalIndex++;
-}
-
-void BlaeckTCP::addSignal(PGM_P const *signalNameTable, int signalNameIndex, unsigned int *value)
-{
-  Signals[_signalIndex].SignalNameTable = signalNameTable;
-  Signals[_signalIndex].SignalNameIndex = signalNameIndex;
-  Signals[_signalIndex].UseFlashSignalName = true;
-  Signals[_signalIndex].DataType = Blaeck_uint;
-  Signals[_signalIndex].Address = value;
-  _signalIndex++;
-}
-
-void BlaeckTCP::addSignal(PGM_P const *signalNameTable, int signalNameIndex, long *value)
-{
-  Signals[_signalIndex].SignalNameTable = signalNameTable;
-  Signals[_signalIndex].SignalNameIndex = signalNameIndex;
-  Signals[_signalIndex].UseFlashSignalName = true;
-  Signals[_signalIndex].DataType = Blaeck_long;
-  Signals[_signalIndex].Address = value;
-  _signalIndex++;
-}
-
-void BlaeckTCP::addSignal(PGM_P const *signalNameTable, int signalNameIndex, unsigned long *value)
-{
-  Signals[_signalIndex].SignalNameTable = signalNameTable;
-  Signals[_signalIndex].SignalNameIndex = signalNameIndex;
-  Signals[_signalIndex].UseFlashSignalName = true;
-  Signals[_signalIndex].DataType = Blaeck_ulong;
-  Signals[_signalIndex].Address = value;
-  _signalIndex++;
-}
-
-void BlaeckTCP::addSignal(PGM_P const *signalNameTable, int signalNameIndex, float *value)
-{
-  Signals[_signalIndex].SignalNameTable = signalNameTable;
-  Signals[_signalIndex].SignalNameIndex = signalNameIndex;
-  Signals[_signalIndex].UseFlashSignalName = true;
-  Signals[_signalIndex].DataType = Blaeck_float;
-  Signals[_signalIndex].Address = value;
-  _signalIndex++;
-}
-
-void BlaeckTCP::addSignal(PGM_P const *signalNameTable, int signalNameIndex, double *value)
-{
-  Signals[_signalIndex].SignalNameTable = signalNameTable;
-  Signals[_signalIndex].SignalNameIndex = signalNameIndex;
-  Signals[_signalIndex].UseFlashSignalName = true;
-  Signals[_signalIndex].DataType = Blaeck_float;
-  Signals[_signalIndex].Address = value;
-  _signalIndex++;
-}
-#endif
 
 void BlaeckTCP::read()
 {
@@ -668,25 +538,8 @@ void BlaeckTCP::writeSymbols(unsigned long msg_id, byte i)
     Clients[i].write((byte)0);
 
     Signal signal = Signals[j];
-#ifdef __AVR__
-    if (signal.UseFlashSignalName)
-    {
-      PGM_P progMemString = (PGM_P)pgm_read_ptr(&signal.SignalNameTable[signal.SignalNameIndex]);
-      char buffer[50];
-      // Copy from flash into RAM
-      strcpy_P(buffer, progMemString);
-      Clients[i].print(buffer);
-      Clients[i].print('\0');
-    }
-    else
-    {
-      Clients[i].print(signal.SignalName);
-      Clients[i].print('\0');
-    }
-#else
     Clients[i].print(signal.SignalName);
     Clients[i].print('\0');
-#endif
 
     switch (signal.DataType)
     {